Zinc integration into Pants.
Not a trivial change. There's quite a bit of commonality between
scala_compile.py and java_compile.py that could be factored out,
but it's best to keep that for a separate change, as it's not
quite as simple as it seems.

(sapling split of 0eafed2668dbfd58d7738a0cab60f1090a817241)