Fix -J argument parsing.
zinc fails on -J arguments because it's also keeping them
in args, then deciding they are invalid arguments.  scalac
will accept them if passed along, but zinc is failing of its
own accord.  This was the easy fix; alternatively you can
pass them along, but it's only necessary if you need scala
to be aware of them.