Add a command for a dirty, noncompiling scala repl.
For debugging purposes, it's often useful to be able to open a repl
on the existing binaries, without having to get code into a currently
compilable state. This provides an alternative scala repl command,
"repl_dirty" that doesn't have the compile dependency. If pants generated binaries
for the targets don't exist, this will fail with an error.

(sapling split of fa19f01e5b14e469852de04ed31a7afee5904051)