Change the locking strategy for Goals.
With this change, the "goal" command is now deserialized, so that
goals that do not require serialization can be started without
grabbing the lock.

Individual goals now specify their serialization needs. By default,
all goals are serialized. If a goal does not require serialization
then it's specification, it can declare "serialize=False".

When goals execute, if a goal requires serialization, the Group executor
will acquire the lock before running the goal. When a serialized goal is
complete, if the next goal in sequence is also serialized, the lock is not
released.

(sapling split of 074cbc304549a4d9b34dd273eea43d3e77b64a67)