Fix zinc rebase implementation.
The old implementation returned None if the source file didn't
actually exist on the filesystem, and also if there was no matching
rebaser for the source file.

Since we only want to rebase files underneath the given bases, and
we don't need to support relativizing paths with ../ etc., we use an
efficient and naive prefix-matching algorithm, instead of analyzing
path segments.

Conflicts:
	src/main/scala/com/typesafe/zinc/SbtAnalysis.scala