Allow pants html report assets to be embedded in the pex.
This is so we don't have to have a copy of all the css, js, templates
etc. in every workspace's build-support.

This still supports overriding this with explicit paths to assets
and templates, in case you want to customize.

Auditors: pl

(sapling split of bc9461f7973f23b852bf57216b7fd67f4f94416d)