Add a transitional pants-packaged target graph.
This adds a set of targets that form a monolithic pants to bridge
past the period where we refactor code and target cycles.  It also
adds a rudimentary test that the pants-packaged target graph
continues to work for publishing pants.

(sapling split of 16c4717903e0d8f9bcad5f20bd58062f40469902)