Kill pants.bootstrap and move its functionality to pants.
There is no reason to ever run pants in the pants repo
from an always-old released versio, so kill pants in its
prior form and move the pants.bootstrap functionality there.

Fixup pants to use requirement.txt files to control 3rdparty
deps.  This allows the BUILDs and the bootstrap to say in
sync with no effort.