Use Scala 2.11.0-RC3 in scriptit tests.
Switch to Scala 2.11.0-RC3 running scriptit tests. Also, update
scala/2.11/jars/test to use full version number for Scala 2.11.0-RC3.
The reason we need to switch to full version is change in sbt's definition
of binaryVersion that scriptit uses. In sbt 0.13 all Scala milestones and
release candidates are considered to be binary incompatible hence
binaryVersion is a full version number.

This has been missed during upgrade to sbt 0.13 of zinc.