Don't add internal jars to the classpath in a bundle jar.
Those classes are already in the jar itself.

It's actually harmful to add those jars to the classpath, due to
the following OS X issue:

The classloader keeps an open handle to each jar on the classpath.
So if you have N targets then you consume N file descriptors. And
we have over 1000 targets in some binaries.

However Java on OS X cannot stat sockets with fds > 1024. This is
probably a bug, but we're stuck with it. So we end up in the
weird situation of the size of our classpath preventing us from
using network sockets...