Remove ability to configure workdir locations.
There's no need to be able to override the location of task workdirs,
and this allows us to be rid of a bunch of flags and pants.ini keys.

After this change we compute a unique workdir per task instance, based
on the phase/goal it's registered under, and plumb that through
the task's __init__(). That did require changing __init__'s signature
on all tasks, but it made for more straightforward code.

While at it, I also fixed a bunch of superclass __init__ invocations
to use super().

This change affects many files (every task at least has its
__init__() signature changed) but the changes are mostly tiny.