Allow old- and new-style resources on python targets.
New-style (specs of Resurces targets) are provided via the resource_targets= arg.
Old-style (lists of files, filesets) are wrapped in a synthetic target.

This allows a smooth transition from old to new-style resources.

If and when all codebases have switched to resource_targets=, and we get rid of
old-style resource support entirely, we can repurpose the resources= arg
to mean the same as resource_targets=.