Only run tests once each...
This was due to the regular target dep-following logic being duplicated
by legacy python_test_suite logic.

This change gets rid of python_test_suite (or rather, makes it
an alias of the 'dependencies' target, to not break existing BUILD files).

We're now back to taking ~2m20s to run all the pants tests.