Use a temporary file as the source of 'y' on stdin for the publish IT.
The prior method of pumping yes to stdin of `./pants goal publish`
did not work under travis-ci.  This method does and it also works
when un-suppressing pytest output as well.

Fixup python integration tests under ci.

Release the lock for the pytest task as is done for the junit task
and other code-run tasks (run, repl).

Convert ci to use the pytest task (test goal) and eliminate dup
runs of python tests by restricting the jvm tests to just the
jvm targets.

This change also adds pass-through args for pytest to aid debugging
tests and otherwise running them.  When using this to un-suppress
pytest output, a second form of test blocking was uncovered with
the stdin 'yes' plumbing not working properly.  As a result this
change tweaks the IT in question to use a more robust method of stdin
piping to the pants subprocess.