Added support for including .proto files in remote .jars on the --proto_path when compiling with protoc.
This is done by way of a new 'imports' parameter added to the java_protobuf_library target, which accepts jar dependencies and jar library targets as inputs. The syntax for this is:

java_protobuf_library(name='foo',
  imports = [
    jar(org='foo.com', name='bar', rev='1.2.3'), jar(org='bar.com', name='foo', rev='3.2.1'),
    '3rdparty:some-jar-library-containing-protos',
  ],
  ...,
)

An extra ivy step called 'ivy-imports' (rather than ivy-resolve) has been added to accomodate this. It occurs after bootstrap and before codegen.