Regression in handling an external dep containing a unicode character classname.
I chose to make the dependency on cucumber-java declared right in the BUILD file and not in 3rdParty/dependencies because I did not think another class would need to reference this library and it isn't needed by the rest of pants, just the example.

Testing Done:
In our case, the error was occuring if you ran 'goal binary' or 'goal bundle' as it ran the 'dups' task.

Added an example project that depends on 'cucumber-java' which has non-ascii classnames.

Added a unit test that depends on it, (but this would not catch the problem we saw because it was in the duplicate detection task)

Added a test to test_duplicate_detection.py to make sure unicode names are properly handled in this step.

Bugs closed: 207

Reviewed at https://rbcommons.com/s/twitter/r/494/