Make jvm and python binary running play nicely with each other.
Previously PythonRun threw an exception if it saw a jvm binary target
on the cmd line. Since it happens to run first, that prevents ever
running a jvm target...

TODO: It would be nice if there was a more principled way for
tasks to coordinate claiming targets, so that we can error out
if the task we're trying to run is neither a jvm_binary nor a
python_binary.