Update to the target_sdk(19) launcher icon
This is from the Android SDK and is released under the Apache license.
Perhaps it would be desirable to swap in a Pants icon.

I am hardcoding the hdpi icon, since I don't think it makes much sense
for pants to be carrying around a bunch of varying ocons resolutions
for such a small example app.