dev docs: talk about Travis-CI in one place, not two. Pull requests trigger CI
Mateo was surprised to find out a pull request gets Travis-CI's attention. I'd like to tease him for that, but, uhm, I didn't know it either. So document it. Once it's documented, I can tease the *next* nerd to hit this.

While I was in there, saw an old claim that pants_tests:all ran all the tests. But now our integration tests are separate. So tweaked that wording.

We had two places that talked about running all the tests (ci.sh/Travis). I was too lazy to add this pull-request blurb to both of 'em, so I consolidated. Now one of them is a hyperlink to the other.

Bugs closed: 267

Reviewed at https://rbcommons.com/s/twitter/r/597/