Raise an exception when the package is missing a template
pkgutils.get_data returns None when it fails to find the requested file. Prior to this change, we would just pass the None on to pystache to render and it would give a confusing message about not being able to find "none_type.mustache". This change checks for the None and raises an Error with the name of the missing template.

I didn't see an obvious existing Error type to use, so I created a new one. Let me know if there's an existing Error I should be using instead.

Testing Done:
ran unittests.

Reviewed at https://rbcommons.com/s/twitter/r/608/