Caching and checking of sdk and build tools working
The dictionary lookup is flawed. It is checking a tuple, but
we actually don't care about structure. Some sort of looking
into the sets that are _validated_sdks and _validated_build_tools
is way way better.

Needs test coverage Y(*.*)Y  (that ascii means I am in trouble
and may run)