WIP: Adding AaptGen task to android backend
Aapt is the codegen tool for Android builds. It spiders downward from
a given directory looking for resources. It creates a R.java, which
allows apps to reference the resources from their java code.
This class is a subclass of codegen, but is installed in the Android
backend.

Current status: './pants goal gen $ANDROID_TARGET' is green.