Aapt_gen task now executing when pointed at sample app
Best day of my life. I tried to get this to happen almost
4 weeks ago. A month worth of integration tests on the
preceding code, here we are.

Now to debug, cause it ain't working. But finding out that the
genlangs function was the problem was a triumph!(lol)

I am afraid that the is_forced is unnecessary, but it was
the only way I could get the targets in gentargets_bylang{}