Return 'manifest' as an android_target param
This allows varible locations for AndroidManifest.xml, and
will more easily support maven-style layouts. This commit
also raises an exception if the manifest is missing, which
we need since we are relying on the manifest to fill out
the target definition.

I added a TODO to support filenames for the manifest other than
AndroidManifest.xml. I believe that allowing the user to type in
their own file and location implies that we support that behavior
already. But we don't (yet).