Add CLI override for target_sdk and build_tools_version
This class will keep growing as our Android functionality does.
Aapt does an awful lot of the android build pipeline.

For now, this allows overriding of the build file and manifest. But
with future tasks, we might need a way to expose mutating
the target itself.

I guess it depends, should changing the target_sdk take affect
globally, or do we want to allow them to mix and match (and therefore
demand they manually add a bunch of other flags if they want a universal
sdk bump without editing the manifest.)