Simplify the JvmTarget resources property.
Resources are just another form of dependency as demonstrated
by the traversable_dependency_specs implementation.  They all
get resolved as dependencies during BUILD parsing so we just need
to look at our direct dependencies that are Resources target to
implement the resources property.  This has the side-benefit that the
standard dependency injection primitives for stitching synthetic targets
into a BUILD graph can be used to stitch in new Resources dependencies.