Further reduce the tool retrieval pipeline
This solves several problems:
  - Multiple classes (and growing) having separate "aapt_tool"
    or "android_jar_tool" methods, with individual APIs.

  - Convuluted validation pipeline, with failure coming
    several methods down the pipe.

With this version, tools are only resolved immediately before they
are invoked. So only a class using "aapt_tool" has an "aapt_tool"
method, and if several classes need the method, use inheritance.

The code is cleaner as well. This works, but more changes coming
to be more lazy with the sdk_path resolve.