Made a small change to stop problems we were getting compiling our annotation processors.
This is related to the pants-devel discussion 'Nondeterministic Phase Execution Order'. Briefly:
when the resources phase happens to precede the compile phase, resources in .pants.d  pollute the
jmake classpath when compiling some of Square's internal annotation processors. This makes jmake
quite irate, so this code stops that from happening.

This fix is not terribly elegant, as it introduces a bogus dependency for resources on compile.
Perhaps this problem will go away when a formal classpath object is created.