Made a small change to stop problems we were getting compiling our annotation processors.
This is related to the pants-devel discussion 'Nondeterministic Phase Execution Order'. Briefly:
when the resources phase happens to precede the compile phase, resources in .pants.d  pollute the
jmake classpath when compiling some of Square's internal annotation processors. This makes jmake
quite irate, so this code stops that from happening.

This fix is not terribly elegant, in that it doesn't address the underlying problem of resources
getting on the classpath in the first place. Perhaps this problem will go away when a formal
classpath object is created.

Testing Done:
Code which didn't compile now compiles.

Ran test tests/python/pants_test:: and nothing fails that doesn't also fail on the master branch.

Bugs closed: 373

Reviewed at https://rbcommons.com/s/twitter/r/708/