Update android target definitions
This is a follow-up based on the comments to merged CR #668.
This can be safely merged without relying on the other two
open reviews.

This gets rid of a couple of the "sensible defaults". I believe
in them as useful, but they have could proven confusing for anyone
with a non-traditional (or just maven-style) layout.

This requires all android targets to explicitly define a "manifest"
field. Android_resource targets must further define a
"resource_dir". I named it "resource_dir" instead of "resources"
to contrast with what pants expects elsewhere: the tool that
processes android resources expects a directory as input, not
a list of files. The aapt tool spiders down the dir, so the
user must specify which directory holds those resources.

Somewhere down the line, I would like to support common layouts
by default. But the way it was before this commit, it implied
that we supported things that we do not.

Testing Done:
local ci.sh passes, android targets can be processed.

Travis looks to be down. [0]

[0] https://travis-ci.org/pantsbuild/pants/builds/30210018

Reviewed at https://rbcommons.com/s/twitter/r/700/