Added a --spec-exclude command line flag to exclude specs by regular expression.
Multiple expressions can be specified by setting the flag more than once. This is in the same
vein as Eric's earlier work on excluding specs with the ^ expression, but this is more powerful
in that it (a) uses regexes, and (b) is an actual flag, and can therefore be specified in
.pants.rc.

Prompted by the need to exclude some automatically generated targets we have internally.

Most of the diff is test code; actual changes are pretty small.

Testing Done:
Wrote tests in pants_test/commands/test_spec_exclude_integration.py (which pass).

Bugs closed: 391

Reviewed at https://rbcommons.com/s/twitter/r/747/