Create a mixin for "quiet tasks".
This separates the two concerns formerly handled by ConsoleTask:
being quiet, and being a task whose primary job is to print to the console.

Now you mix in QuietTaskMixin for the first, and extend ConsoleTask
(which itself mixes in QuietTaskMixin) for the second.

This rids us of silliness such as the clean-all task having a flag
to specify its output separator.