Properly order resources for pants goal test and pants goal run.
Fixed a problem with goal run and test for resources being put on the classpath in the
wrong order when multiple resource paths contain the same resource path.

Added a project with resources that collide and an integration test to prove
that resources are bundled and added to the classpath in the right order.

Note that the JVM tests will still fail if all run together with

   pants goal test testprojects/maven_layout/resource_collision::

because pants runs all the tests in the same junit test runner instance.