Gathered interesting classes in a set for dexing
This looks wrong to me still. I could filter the 'classes_by_target'
with Android targets, but how to show that the classes that originate
from aapt_gen (R.class) and the ones frmo the AndroidBinary are part of the
same package.

Going to implement the dexing now, and kick the target grouping down
the road. Hopefully it will become clearer while I work, or else I
can ask the list.