Clean up and doc DxCompile
This class is ready except for finding if there is an easy
way to pass VM options to dx. If not then it needs to have
the setup_parser completed.

One lingering question is where to put output. I am tempted
to move the output from aapt and DxCompile to the dist folder,
since I will need it acessbile for aapt_package. We will see if
it is needed when we work on that class.