Update front-facing help and error messages for Android targets/tasks
Small improvements to an exception message and a CLI flag.
The missing manifest error in android_target didn't give
any hints as to where the bad BUILD file entry was pointing.

And the sdk-path CLI flag for android_task wasn't clear about what it did.
It could easily have been interpreted as forcing a platform
version (i.e. SDK platform 17) as opposed to what it
actually is meant to do, which is point to an entire
Android SDK installation somewhere on the filesystem.

The use case is to force a custom SDK over the AOSP one on
the machines path. Lots of OEM customize the framework and
might have multiple SDKs on the fs.

Testing Done:
CI passes.

Travis passes [0].

[0] https://travis-ci.org/pantsbuild/pants/builds/31784570

Reviewed at https://rbcommons.com/s/twitter/r/837/