Move CodeGen.package_path into AaptTask base task
No one else is using it, and I needed it outside of codegen.
If someone thinks we should move it even higher so that
codegen can still get it, I am all ears. But even in codegen, it
seemed a little out of place since I don't believe all codegen
targets even have direct concept of package or manifest.