Add package_name to Android targets
This attribute is filled by the manifest parser. If unable
to find a proper package name, this falls back silently to
the target's name. This attribute is used to name to output
apk. Our sample app's android:name is
com.pants.example.hello.HelloWorld, which now outputs an apk called
HelloWorld-unsigned.apk. If no android:name was found, we would
see a package named hello-unisgned.apk.