Fix bad BUILD file and failed test for AaptGen tests
This is off, has been for a little while. A change broke
the test, but it had an inconsistent BUILD file so that
test was never getting called by the CI.

I have a big refactor of the entire Aapt pipeline
coming up with the AaptBuild task, which I am writing tests
for now. But this should go in soon, because the this test
does fail if the user calls something like
'goal test tests::', etc.