AndroidIntegrationTest class added
This class is passed a list of TOOL paths and verifies them against
the ANDROID_HOME environmental variable.

Right now I have all the constants defined in AndroidIntegration,
as in BUILD_TOOLS_VERSION and TARGET_SDK. That is to keep
from defining them in 100 places, but I could be swayed into
defining in the subclasses individually. I chose the super class
because all of the int-tests are using the same sample target, so
any change in the sample app would necessitate changing these values
for every class that defined them.