Add a BUILD.debug to hold keystore info
This is perfectly servicable, but the only thing stopping
someone from putting their release credentials in the vcs-indexed
BUILD.debug is a warning. That is non-ideal.

I would like to add some protections before we add release
builds. Ideas include, double-checking that the 'release' target is
in a file named BUILD.release, and verifiying that is in gitignore.
Perhaps use textfile based config files, and mandate the release config
be outside pants buildroot.

Still thinking.