Fix bad BUILD file and failed test for AaptGen tests
This is independent of any other of my reviews currently up.

This test was broken, had been for a little while. A change broke
the test but it had an inconsistent BUILD file so that
test was never getting called by the CI.

I have a somwhat large refactor of the entire Aapt pipeline
coming up with the AaptBuild task, which I am writing tests
for now. But I wanted to get this in soon, because the this test
will fail if the user calls something like
'./pants goal test tests::', etc.

Testing Done:
Verified that all android tests are being called by CI.
CI passes.
Travis passes.

Reviewed at https://rbcommons.com/s/twitter/r/858/